Bug A - Make KeyframeEffect weak-referenceable
We want to make KeyframeEffect replaceable by a more compact (and
ultimately coalesceable) form such that the original KeyframeEffect, if
not referenced by script, can be garbage collected. So long as the
original object exists, however, we need to make sure it is kept
up-to-date. As a result we make it weak-referenceable.