Bug A - Add an internal-use-only Fixed timing function
When we have a filling animation that fills between two keyframes,  we
need to remember the portion between the keyframes so we can interpolate
at the appropriate point.

We could use a linear timing function and adjust the keyframe offsets to
ensure we end up interpolating at the right point, but that won't help
us when the portion is less than zero or greater than one. Furthermore,
tweaking the keyframe offsets is complicated. Instead we just add a new
timing function type that takes a single parameter--a progress
value--whose output is always that value.