Bug A - Add an additional animation index so that linked effects can sort according to the effects they represent
There are alternative approaches we could try such as having compacted
animations and source animations swap indices or copy them however this
approach of adding a secondary index allows us to apply it to
FillAnimations too when we introduce them in a later bug. Furthermore,
the other approaches tend to fall down in the face of events which may
be dispatched by an animation prior to being replaced but then sorted
after being replaced.