Bug A - Recompact an animation when it changes target
This will change the EffectSet it needs to be registerd with. Rather
than trying to get the compacted effect to update its target, it's
easier to just re-wrap the element.

From a spec point-of-view if a FillAnimation represents two filling
KeyframeEffects and one of them is retargeted at a different element,
then the identity of the FillAnimation will change and we should
re-generate it. For consistency, it makes sense to do the same for the
single effect case too.