Bug A - Recompact a KeyframeEffect when it changes animation
When an effect is assigned to a different animation it may no longer be
eligible for compacting (e.g. it might be assigned to an animation with
a non-monotonically increasing timeline, or an animation with
a different current time such that it is no longer filling). As a result
we need to restore the animation even though in some cases we may
subsequently compact it again.

We try to contain as much of the code relating to compacting in
KeyframeEffect itself hence the changes here are mostly to
KeyframeEffect. However, in order to be able to restore the original
Animation by using CompactAnimationUtils::RestoreAnimation we need to
have access to the animation prior to it being disconnected from the
effect so this patch introduces an extra BeforeSetAnimation function for
that purpose.