Bug B - Make CompactFillEffects and FillEffects cancel each other
For example if the user cancels a FillAnimation, this needs to cancel
any CompactFillEffects it represents since it is these
CompactFillEffects that actually influence style.

Likewise, if a CompactFillEffect is canceled via a linked effect, that
is, via the original Animation from which is was created, then any
FillAnimations that refer to the CompactFillEffect will need to be
canceled as well.

This second case will be further refined in the next patch in this
series so that it does not subsequently trigger the first case too.