Bug B - Implement FillEffect::GetProperties
This is just the combination of the properties of all source effects.

We don't do any sorting of these properties. Furthermore, the result of
calling this will expose whether or not effects have been combined
internally.  However, those limitations that seem acceptable for now
since this is not Web-facing. We can refine the result of this function
later according to the needs of DevTools (or do any sorting and
simplification on the DevTools side) if needed.