Bug C - Track the global animation list using a linked list
This is so that we will be able to determine if two effects are eligible
for combining. If there exists some other animation that occurs in
between the two effects' associated animations in the global animation
list, even if it is idle, or does not currently target the same element,
we still should not combine the effects since it's possible that this
interleaving animation could be modified in a way that means we need to
separate out the effects again.

Hopefully this will also be useful if we later introduce an API for
mutating the composite order of animations (e.g.
compositeBefore/compositeAfter).